+++ lib/scalingo/client.rb
@@ -7,23 +7,38 @@ module Scalingo
   class Client < CoreClient
     ## API clients
     def auth
-      @auth ||= Auth.new("https://auth.scalingo.com/v1", self)
+      @auth ||= Auth.new(
+        "https://auth.scalingo.com/v1",
+        scalingo: self,
+      )
     end
 
     def billing
-      @billing ||= Billing.new("https://cashmachine.scalingo.com", self)
+      @billing ||= Billing.new(
+        "https://cashmachine.scalingo.com",
+        scalingo: self,
+      )
     end
 
     def agora_fr1
-      @agora_fr1 ||= Regional.new("https://api.agora-fr1.scalingo.com/v1", self)
+      @agora_fr1 ||= Regional.new(
+        "https://api.agora-fr1.scalingo.com/v1",
+        scalingo: self,
+      )
     end
 
     def osc_fr1
-      @osc_fr1 ||= Regional.new("https://api.osc-fr1.scalingo.com/v1", self)
+      @osc_fr1 ||= Regional.new(
+        "https://api.osc-fr1.scalingo.com/v1",
+        scalingo: self,
+      )
     end
 
     def osc_secnum_fr1
-      @osc_secnum_fr1 ||= Regional.new("https://api.osc-secnum-fr1.scalingo.com/v1", self)
+      @osc_secnum_fr1 ||= Regional.new(
+        "https://api.osc-secnum-fr1.scalingo.com/v1",
+        scalingo: self,
+      )
     end
   end
 end