+++ spec/scalingo/auth_spec.rb
@@ -1,7 +1,7 @@
 require "spec_helper"
 
 RSpec.describe Scalingo::Auth do
-  subject { described_class.new(:client, :url) }
+  subject { described_class.new("url") }
 
   %w[keys scm_integrations tokens two_factor_auth user].each do |section|
     it "handles requests for #{section}" do